Добавлен интерактор для работы с сетью. Создана разметка для элемента списка(вместо создания в коде)

+++ app/src/main/java/com/milnest/tasklist/repository/DBRepository.kt
@@ -115,7 +115,7 @@ object DBRepository {
                 arrayOf(id.toString())).toLong()
     }
 
-    fun deleteTask(id: Int) {
+    fun deleteTask(id: Long) {
         db.delete(TaskDatabaseHelper.TABLE, TaskDatabaseHelper.COLUMN_ID +
                 " =?", arrayOf(id.toString()))
     }
